current affairs

+++ src/helpers/blogApi.js
@@ -1,6 +1,6 @@
 import httpClient from '@/helpers/httpClient';
 
-export const getBlogs = () => httpClient.get('/api/blogs');
-export const createBlog = (data) => httpClient.post('/api/blogs', data);
-export const updateBlog = (id, data) => httpClient.put(`/api/blogs/${id}`, data);
-export const deleteBlog = (id) => httpClient.delete(`/api/blogs/${id}`);
+export const getBlogs = () => httpClient.get('/api/admin/blogs');
+export const createBlog = (data) => httpClient.post('/api/admin/blogs', data);
+export const updateBlog = (id, data) => httpClient.put(`/api/admin/blogs/${id}`, data);
+export const deleteBlog = (id) => httpClient.delete(`/api/admin/blogs/${id}`);
